require unbroken chain to primary
in order to have a meaningful ground truth labelling, we need an unbroken chain from every particle back to its primary. in the absence of an unbroken chain, we may lack vital information. for that reason, we have to throw an error if we find a particle we need is missing from the truth list.